feat: move invoices to payments page
