+++ components/Transaction/transaction.module.css
@@ -351,7 +351,7 @@ body[data-theme='dark'] .edit_invoice img {
   filter: invert(1);
 }
 
-body[data-theme='dark'] .see_invoice img {
+body[data-theme='dark'] .view_invoice img {
   filter: invert(1);
 }
 
@@ -365,7 +365,7 @@ body[data-theme='dark'] .see_invoice img {
     brightness(100%) contrast(102%);
 }
 
-.see_invoice_ctn:hover .see_invoice {
+.view_invoice_ctn:hover .view_invoice {
   filter: invert(60%) sepia(14%) saturate(5479%) hue-rotate(195deg)
     brightness(100%) contrast(102%);
 }
