+++ pages/payments/payments.module.css
@@ -59,7 +59,7 @@ body[data-theme='dark'] .show_filters_button img {
 
 .filter_button:active {
     transform: scale(0.95);
-  }
+}
 
 .active {
     background-color: var(--accent-color);
@@ -80,10 +80,250 @@ body[data-theme='dark'] .show_filters_button img {
 
 .wallet_label {
     margin-top: 20px !important;
- }
+}
 
- .filter_btns {
+.filter_btns {
     display: flex;
     align-items: center;
     gap: 10px;
- }
+}
+
+.form_ctn_outer {
+    background-color: rgba(255, 255, 255, 0.3);
+    backdrop-filter: blur(5px);
+    position: fixed;
+    top: 0;
+    left: 0px;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    z-index: 99999;
+}
+
+.form_ctn_inner {
+    width: 100%;
+    max-width: 600px;
+    align-self: middle;
+}
+
+.form_ctn_inner h4 {
+    margin: 0;
+    margin-left: 5px;
+    margin-bottom: 5px;
+}
+
+.form_ctn {
+    background-color: var(--secondary-bg-color);
+    width: 100%;
+    padding: 25px;
+    border-radius: 10px;
+    border: 1px solid var(--secondary-text-color);
+    overflow-y: auto;
+    max-height: 90vh;
+}
+
+body[data-theme='dark'] .form_ctn_outer {
+    background-color: rgba(0, 0, 0, 0.3);
+}
+
+.form_ctn p {
+    margin-top: 0;
+}
+
+.form_ctn form {
+    display: flex;
+    flex-direction: column;
+}
+
+.form_ctn label {
+    margin-bottom: 5px;
+}
+
+body[data-theme='dark'] .form_ctn select {
+    color: #fff;
+}
+
+.form_ctn select {
+    background-color: var(--secondary-bg-color);
+}
+
+.form_ctn input,
+.form_ctn textarea {
+    border: 1px solid var(--secondary-text-color) !important;
+    width: 100%;
+}
+
+body[data-theme='dark'] .form_ctn input,
+body[data-theme='dark'] .form_ctn textarea,
+body[data-theme='dark'] .form_ctn select {
+    background-color: var(--primary-bg-color);
+    border-color: #898EA4 !important;
+}
+
+body[data-theme='dark'] .form_ctn select:not([multiple]) {
+    background-image: linear-gradient(45deg,transparent 49%,#fff 51%),linear-gradient(135deg,#fff 51%,transparent 49%);
+}
+
+.labelMargin {
+    margin-top: 10px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.form_ctn button:last-child {
+    margin-right: 0px;
+    box-shadow: none;
+}
+
+.form_ctn .delete_btn {
+    background-color: unset;
+    border: none;
+    margin-right: 0px;
+    margin-left: 10px;
+    color: var(--primary-text-color);
+    border-radius: 0px;
+    padding-left: 0px;
+    padding-right: 0px;
+    opacity: 0.6;
+    position: relative;
+    transition: all ease-in-out 100ms;
+}
+
+.form_ctn .delete_btn:hover {
+    color: red;
+    opacity: 1;
+}
+
+.form_ctn .delete_btn div {
+    width: 15px;
+    border-radius: 0px !important;
+    position: absolute;
+    right: 0px;
+    top: 12px;
+    opacity: -10;
+    transition: all ease-in-out 100ms;
+}
+
+.form_ctn .delete_btn:hover div {
+    right: -18px;
+    opacity: 1;
+}
+
+.form_ctn .delete_confirm_btn {
+    padding-left: 20px;
+    padding-right: 20px;
+    border: 1px solid var(--primary-text-color);
+    background-color: var(--primary-bg-color);
+    color: red;
+    transition: all ease-in-out 100ms;
+    font-weight: 600;
+}
+
+.delete_button_form_ctn {
+    word-break: normal;
+}
+
+.form_ctn .delete_confirm_btn:hover {
+    border: 1px solid var(--primary-text-color);
+    background-color: rgb(149, 0, 0);
+    color: var(--secondary-bg-color);
+}
+
+.tooltiptext {
+    visibility: hidden;
+    opacity: 0;
+    background-color: var(--secondary-bg-color);
+    color: var(--primary-text-color);
+    text-align: center;
+    padding: 2px 0px;
+    border-radius: 5px;
+    position: absolute;
+    z-index: 1;
+    bottom: -30px;
+    width: 80px;
+    right: -5px;
+    flex-basis: 1;
+    font-size: 12px;
+    z-index: 99999999;
+    box-sizing: border-box;
+    transition: all 100ms ease-in-out;
+}
+
+.tooltiptext::before {
+    content: '';
+    position: absolute;
+    left: 50%;
+    width: 0;
+    height: 0;
+    border-left: 7px solid transparent;
+    border-right: 7px solid transparent;
+    transform: translate(-50%, -9px);
+    border-bottom: 7px solid var(--secondary-bg-color);
+}
+
+.create_invoice:hover .tooltiptext {
+    visibility: visible;
+    bottom: -35px;
+    opacity: 1;
+}
+
+.create_invoice_ctn {
+    position: relative;
+    position: sticky;
+    bottom: 0;
+    float: right;
+}
+
+.create_invoice {
+    background-color: var(--secondary-bg-color);
+    font-size: 44px;
+    display: inline-block;
+    border-radius: 100px;
+    padding: 20px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    transition: all ease-in-out 100ms;
+    position: relative;
+    border: 1px solid var(--secondary-text-color);
+}
+
+.invoice_view_item {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 10px;
+    border-bottom: 1px solid var(--primary-text-color);
+}
+
+body[data-theme='dark'] .create_invoice img {
+    filter: invert(1);
+}
+
+body[data-theme='dark'] .edit_invoice img {
+    filter: invert(1);
+}
+
+body[data-theme='dark'] .view_invoice img {
+    filter: invert(1);
+}
+
+.create_invoice_ctn:hover .create_invoice {
+    filter: invert(60%) sepia(14%) saturate(5479%) hue-rotate(195deg)
+    brightness(100%) contrast(102%);
+}
+ 
+.edit_invoice_ctn:hover .edit_invoice{
+    filter: invert(60%) sepia(14%) saturate(5479%) hue-rotate(195deg)
+    brightness(100%) contrast(102%);
+}
+
+.view_invoice_ctn:hover .view_invoice {
+    filter: invert(60%) sepia(14%) saturate(5479%) hue-rotate(195deg)
+        brightness(100%) contrast(102%);
+}
+
