+++ services/transactionService.ts
@@ -663,13 +663,31 @@ export async function getPaymentsByUserIdOrderedByButtonName (
           'priceValue', pb.value,
           'quoteId', pb.quoteId
         )
-      ) AS prices
+      ) AS prices,
+      JSON_ARRAYAGG(
+        IF(i.id IS NOT NULL,
+          JSON_OBJECT(
+            'id', i.id,
+            'invoiceNumber', i.invoiceNumber,
+            'amount', i.amount,
+            'description', i.description,
+            'recipientName', i.recipientName,
+            'recipientAddress', i.recipientAddress,
+            'customerName', i.customerName,
+            'customerAddress', i.customerAddress,
+            'createdAt', i.createdAt,
+            'updatedAt', i.updatedAt
+          ),
+          NULL
+        )
+      ) AS invoices
     FROM \`Transaction\` t
     INNER JOIN \`Address\` a ON t.\`addressId\` = a.\`id\`
     INNER JOIN \`AddressesOnButtons\` ab ON a.\`id\` = ab.\`addressId\`
     INNER JOIN \`Paybutton\` p ON ab.\`paybuttonId\` = p.\`id\`
     LEFT JOIN \`PricesOnTransactions\` pt ON t.\`id\` = pt.\`transactionId\`
     LEFT JOIN \`Price\` pb ON pt.\`priceId\` = pb.\`id\`
+    LEFT JOIN \`Invoice\` i ON i.\`transactionId\` = t.\`id\`
     WHERE t.\`amount\` > 0
     AND EXISTS (
       SELECT 1
@@ -708,12 +726,14 @@ export async function getPaymentsByUserIdOrderedByButtonName (
     })
     if (tx.amount > 0) {
       payments.push({
+        id: tx.id,
         amount: tx.amount,
         timestamp: tx.timestamp,
         values: ret,
         networkId: tx.networkId,
         hash: tx.hash,
-        buttonDisplayDataList
+        buttonDisplayDataList,
+        invoices: JSON.parse(tx.invoices)
       })
     }
   })
@@ -783,7 +803,7 @@ export async function fetchAllPaymentsByUserIdWithPagination (
 
   const transactions = await prisma.transaction.findMany({
     where,
-    include: includePaybuttonsAndPrices,
+    include: includePaybuttonsAndPricesAndInvoices,
     orderBy: orderByQuery,
     skip: page * Number(pageSize),
     take: Number(pageSize)
